Converting tensor to double automatically

+++ spng/src/TorchToTensor.cxx
@@ -36,7 +36,7 @@ bool TorchToTensor::operator()(const input_pointer& torchset_in, output_pointer&
     ITensor::vector tensor_vec;
     for (const auto & tensor_ptr : (*torchset_in->tensors())) {
         //Make sure to move to CPU
-        auto tensor_clone = tensor_ptr->tensor().clone().to(torch::kCPU).contiguous();
+        auto tensor_clone = tensor_ptr->tensor().clone().to(torch::kCPU).to(torch::kDouble).contiguous();
         // const auto shape = ten->shape();
     
         std::vector<double> as_vec(
